+++ libs/feature-rooms/src/lib/message-list/simple-message-list/simple-message-list.component.spec.ts
@@ -33,6 +33,7 @@ function msg(
     caption: null,
     captionHtml: null,
     readReceipts: [],
+    poll: null,
   };
 }
 
@@ -104,6 +105,7 @@ describe('SimpleMessageListComponent', () => {
             caption: null,
             captionHtml: null,
             readReceipts: [],
+            poll: null,
           },
         ],
       },