+++ libs/feature-rooms/src/lib/message-list/virtual-message-list/virtual-message-list.component.spec.ts
@@ -32,6 +32,7 @@ function msg(
     caption: null,
     captionHtml: null,
     readReceipts: [],
+    poll: null,
   };
 }