+++ libs/util-matrix/src/index.ts
@@ -6,6 +6,7 @@ export * from './lib/presence';
 export * from './lib/message-content';
 export * from './lib/typing';
 export * from './lib/matrix-to';
+export * from './lib/poll';
 export * from './lib/transient-errors';
 export * from './lib/password-uia';
 export * from './lib/attachment-crypto';